pyproject.toml
[tool.poetry]
name = "drfcrud-testtask"
version = "0.1.0"
description = ""
authors = ["Oleksii R <88157588+alexsproject@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
